test : Path 요금 계산 테스트 추가
